initial tenants

+++ src/AppBundle/Controller/CompanyController.php
@@ -11,6 +11,7 @@ use AppBundle\Entity\Company;
 use Pagerfanta\Adapter\DoctrineORMAdapter;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Pagerfanta\Pagerfanta;
+use Symfony\Component\Finder\Exception\AccessDeniedException;
 
 /**
  * @Route("/app")
@@ -25,7 +26,7 @@ class CompanyController extends Controller
     {
         $companiesQuery = $this->getDoctrine()
                         ->getRepository('AppBundle:Company')
-                        ->findAllQuery();
+                        ->findAllQuery($this->get('session')->get('tenant'));
 
         $adapter = new DoctrineORMAdapter($companiesQuery);
 
@@ -41,6 +42,11 @@ class CompanyController extends Controller
      */
     public function deleteCompanyAction(Company $company, Request $request)
     {
+        if (!$this->get('app.helper.tenant')->isTenantObjectOwner($company))
+        {
+            throw new AccessDeniedException();
+        }
+
         $em = $this->getDoctrine()->getEntityManager();
         $em->remove($company);
         $em->flush();
@@ -73,6 +79,11 @@ class CompanyController extends Controller
                 continue;
             }
 
+            if (!$this->get('app.helper.tenant')->isTenantObjectOwner($company))
+            {
+                throw new AccessDeniedException();
+            }
+
             $em->remove($company);
             $em->flush();
 
@@ -89,6 +100,10 @@ class CompanyController extends Controller
      */
     public function showCompanyAction(Company $company, Request $request)
     {
+        if (!$this->get('app.helper.tenant')->isTenantObjectOwner($company))
+        {
+            throw new AccessDeniedException();
+        }
 
         return ['company' => $company];
     }
@@ -100,6 +115,11 @@ class CompanyController extends Controller
      */
     public function editCompanyAction(Company $company, Request $request)
     {
+        if (!$this->get('app.helper.tenant')->isTenantObjectOwner($company))
+        {
+            throw new AccessDeniedException();
+        }
+
         $form = $this->get('app.form.company')->setData($company);
         $form->handleRequest($request);
 
@@ -134,7 +154,9 @@ class CompanyController extends Controller
         if ($form->isValid())
         {
             $em = $this->getDoctrine()->getManager();
+            $tenant = $em->getRepository('AppBundle:Tenant')->find($this->get('session')->get('tenant')->getId());
             $company->setUser($user);
+            $company->setTenant($tenant);
             $em->persist($company);
             $em->flush();
